rename decoder module
